# docs/swagger.yaml
definitions:
  items.CreateItems:
    properties:
      amount:
        type: integer
      name:
        type: string
    type: object
  items.Items:
    properties:
      _id:
        type: string
      amount:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /items:
    get:
      consumes:
      - application/json
      description: Retrieve all items from the items collection
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/items.Items'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create a new item in the items collection
      parameters:
      - description: Item to create
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/items.CreateItems'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new item
      tags:
      - items
  /items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an item by its ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete an item
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Get a specific item by its ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/items.Items'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get item by ID
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update an existing item by its ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated item data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/items.Items'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update an item
      tags:
      - items
swagger: "2.0"